+++ processor/src/test/resources/expected/com/example/repository/CompleteRepositoryExampleRepository.java
@@ -11,11 +11,11 @@ import java.util.stream.Stream;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.realityforge.arez.Arez;
 import org.realityforge.arez.Disposable;
 import org.realityforge.arez.Observable;
 import org.realityforge.arez.annotations.Action;
 import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.ObservableRef;
 import org.realityforge.arez.annotations.PreDispose;
 import org.realityforge.braincheck.Guards;
 
@@ -25,9 +25,6 @@ import org.realityforge.braincheck.Guards;
 )
 public class CompleteRepositoryExampleRepository implements CompleteRepositoryExampleBaseRepositoryExtension, CompleteRepositoryExample.FooEx {
   private static final boolean $$arez$$_IMMUTABLE_RESULTS = "true".equals( System.getProperty( "arez.repositories_return_immutables", String.valueOf( System.getProperty( "arez.environment", "production" ).equals( "development" ) ) ) );
-  ;
-
-  private final Observable $$arez$$_observable = Arez.context().createObservable( Arez.context().areNamesEnabled() ? "CompleteRepositoryExampleRepository.entities" : null );
   ;
 
   private final HashMap<Integer, CompleteRepositoryExample> $$arez$$_entities = new HashMap<>();
@@ -51,7 +48,7 @@ public class CompleteRepositoryExampleRepository implements CompleteRepositoryEx
   public CompleteRepositoryExample create(@Nonnull final String packageName, @Nonnull final String name) {
     final Arez_CompleteRepositoryExample entity = new Arez_CompleteRepositoryExample(packageName,name);
     $$arez$$_entities.put( entity.getId(), entity );
-    $$arez$$_observable.reportChanged();
+    getEntitiesObservable().reportChanged();
     return entity;
   }
 
@@ -59,11 +56,11 @@ public class CompleteRepositoryExampleRepository implements CompleteRepositoryEx
   final void preDispose() {
     $$arez$$_entityList.forEach( e -> Disposable.dispose( e ) );
     $$arez$$_entities.clear();
-    $$arez$$_observable.reportChanged();
+    getEntitiesObservable().reportChanged();
   }
 
   public boolean contains(@Nonnull final CompleteRepositoryExample entity) {
-    $$arez$$_observable.reportObserved();
+    getEntitiesObservable().reportObserved();
     return $$arez$$_entities.containsKey( entity.getId() );
   }
 
@@ -72,7 +69,7 @@ public class CompleteRepositoryExampleRepository implements CompleteRepositoryEx
     assert null != entity;
     if ( null != $$arez$$_entities.remove( entity.getId() ) ) {
       Disposable.dispose( entity );
-      $$arez$$_observable.reportChanged();
+      getEntitiesObservable().reportChanged();
     } else {
       Guards.fail( () -> "Called destroy() passing an entity that was not in the repository. Entity: " + entity );
     }
@@ -80,8 +77,13 @@ public class CompleteRepositoryExampleRepository implements CompleteRepositoryEx
 
   @Nullable
   public CompleteRepositoryExample findById(final int id) {
-    $$arez$$_observable.reportObserved();
+    getEntitiesObservable().reportObserved();
     return $$arez$$_entities.get( id );
+  }
+
+  @ObservableRef
+  Observable getEntitiesObservable() {
+    throw new IllegalStateException();
   }
 
   /**
@@ -89,9 +91,11 @@ public class CompleteRepositoryExampleRepository implements CompleteRepositoryEx
    * This collection should not be exposed to the user but may be used be repository extensions when
    * they define custom queries. NOTE: use of this method marks the list as observed.
    */
+  @org.realityforge.arez.annotations.Observable(
+      expectSetter = false
+  )
   @Nonnull
-  protected final Collection<CompleteRepositoryExample> entities() {
-    $$arez$$_observable.reportObserved();
+  protected Collection<CompleteRepositoryExample> entities() {
     return $$arez$$_entityList;
   }